Consolidate tests to controller groups.
