+++ src/SDKs/CognitiveServices/dataPlane/Personalizer/Personalizer.Tests/Personalizer/LogsTests.cs
@@ -1,14 +1,13 @@
-using System;
-using System.IO;
+﻿using Microsoft.Azure.CognitiveServices.Personalizer.Models;
 using Microsoft.Azure.Test.HttpRecorder;
 using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
-using System.Threading.Tasks;
 using Xunit;
-using Microsoft.Azure.CognitiveServices.Personalizer.Models;
+using System;
+using System.Threading.Tasks;
 
 namespace Microsoft.Azure.CognitiveServices.Personalizer.Tests
 {
-    public class GetLogsPropertiesTest : BaseTests
+    public class LogsTests : BaseTests
     {
         [Fact]
         public async Task GetLogsProperties()
@@ -22,6 +21,19 @@ namespace Microsoft.Azure.CognitiveServices.Personalizer.Tests
                 LogsProperties properties = await client.GetLogsPropertiesAsync();
                 Assert.Equal(new DateTime(2019, 04, 13), new DateTime(properties.DateRange.FromProperty.Value.Year, properties.DateRange.FromProperty.Value.Month, properties.DateRange.FromProperty.Value.Day));
                 Assert.Equal(new DateTime(2019, 04, 23), new DateTime(properties.DateRange.To.Value.Year, properties.DateRange.To.Value.Month, properties.DateRange.To.Value.Day));
+            }
+        }
+
+        [Fact]
+        public async Task DeleteLogs()
+        {
+            using (MockContext.Start(this.GetType().FullName))
+            {
+                HttpMockServer.Initialize(this.GetType().FullName, "DeleteLogs");
+
+                IPersonalizerClient client = GetClient(HttpMockServer.CreateInstance());
+
+                await client.DeleteLogsAsync();
             }
         }
     }
