+++ src/SDKs/CognitiveServices/dataPlane/Personalizer/Personalizer.Tests/SessionRecords/Microsoft.Azure.CognitiveServices.Personalizer.Tests.ConfigurationsTests/GetServiceConfiguration.json
